+++ src/lib/packet.c
@@ -57,7 +57,7 @@ gpac_pck_get_stream_time(GstClockTime time,
   // DTS value
   gboolean is_negative = ret < 0;
   if (is_dts && is_negative && !priv->dts_offset_set) {
-    priv->dts_offset = unsigned_time;
+    priv->dts_offset = (gint64)unsigned_time;
     priv->dts_offset_set = TRUE;
   }
